chore: move grpc agent package
